Ignore the log files when they are generated
Create two functions in the common module to get the full path of the
log files (debug log and reporter log) anywhere in the script. It will
avoid duplicate reference to the names of the log files.
Then whitelist the log file(s) generated to avoid reporting warnings.
Note that log files generated from another run will still be reported.