Create new test cases to validate the new feature
Fix the existing tests by adding the arguments "reporter" and
"enable_debug_log" in the stub of argparser (Args()).
Create new test cases to validate the log files are correctly ignored
or reported.

Also improve a bit the code in check_file_whitelist: check if the
file needs to be ignored at the very beginning of the loop since the
the other check (using regex) requires more lines to be executed.